# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sale Management API",
    "version": "1.0.0",
    "description": "API documentation for Sale Management System"
  },
  "servers": [
    { "url": "http://localhost:4000" }
  ],
  "paths": {
    "/api/client-types": {
      "get": { "summary": "Get all client types", "responses": { "200": { "description": "List of client types" } } },
      "post": {
        "summary": "Create a client type",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ClientType" } } } },
        "responses": { "200": { "description": "Client type created" } }
      }
    },
    "/api/client-types/{id}": {
      "get": { "summary": "Get client type by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Client type data" } } },
      "put": { "summary": "Update client type", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ClientType" } } } }, "responses": { "200": { "description": "Client type updated" } } },
      "delete": { "summary": "Delete client type", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Client type deleted" } } }
    },
    "/api/clients": {
      "get": { "summary": "Get all clients", "responses": { "200": { "description": "List of clients" } } },
      "post": { "summary": "Create a client", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Client" } } } }, "responses": { "200": { "description": "Client created" } } }
    },
    "/api/clients/{id}": {
      "get": { "summary": "Get client by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Client data" } } },
      "put": { "summary": "Update client", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Client" } } } }, "responses": { "200": { "description": "Client updated" } } },
      "delete": { "summary": "Delete client", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Client deleted" } } }
    },
...
    "/api/employees": {
      "get": { "summary": "Get all employees", "responses": { "200": { "description": "List of employees" } } },
      "post": { "summary": "Create an employee", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Employee" } } } }, "responses": { "200": { "description": "Employee created" } } }
    },
    "/api/employees/{id}": {
      "get": { "summary": "Get employee by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Employee data" } } },
      "put": { "summary": "Update employee", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Employee" } } } }, "responses": { "200": { "description": "Employee updated" } } },
      "delete": { "summary": "Delete employee", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Employee deleted" } } }
    },
...
    "/api/invoices": {
      "get": { "summary": "Get all invoices", "responses": { "200": { "description": "List of invoices" } } },
      "post": { "summary": "Create an invoice", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invoice" } } } }, "responses": { "200": { "description": "Invoice created" } } }
    },
    "/api/invoices/{id}": {
      "get": { "summary": "Get invoice by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Invoice data" } } },
      "put": { "summary": "Update invoice", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invoice" } } } }, "responses": { "200": { "description": "Invoice updated" } } },
      "delete": { "summary": "Delete invoice", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Invoice deleted" } } }
    },
    "/api/invoice-details": {
      "get": { "summary": "Get all invoice details", "responses": { "200": { "description": "List of invoice details" } } },
      "post": { "summary": "Create invoice detail", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvoiceDetail" } } } }, "responses": { "200": { "description": "Invoice detail created" } } }
    },
    "/api/invoice-details/{invoiceNo}/{productNo}": {
      "put": { "summary": "Update invoice detail", "parameters": [
        { "name": "invoiceNo", "in": "path", "required": true, "schema": { "type": "integer" } },
        { "name": "productNo", "in": "path", "required": true, "schema": { "type": "string" } }
      ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvoiceDetail" } } } }, "responses": { "200": { "description": "Invoice detail updated" } } },
      "delete": { "summary": "Delete invoice detail", "parameters": [
        { "name": "invoiceNo", "in": "path", "required": true, "schema": { "type": "integer" } },
        { "name": "productNo", "in": "path", "required": true, "schema": { "type": "string" } }
      ], "responses": { "200": { "description": "Invoice detail deleted" } } }
    }
  },
  "components": {
    "schemas": {
      "ClientType": { "type": "object", "properties": { "CLIENT_TYPE": { "type": "string" }, "TYPE_NAME": { "type": "string" }, "DISCOUNT_RATE": { "type": "number" }, "REMARKS": { "type": "string" } }, "required": ["CLIENT_TYPE", "TYPE_NAME"] },
      "Client": { "type": "object", "properties": { "ClientName": { "type": "string" }, "Address": { "type": "string" }, "City": { "type": "string" }, "Phone": { "type": "string" }, "CLIENT_TYPE": { "type": "string" }, "Discount": { "type": "number" } }, "required": ["ClientName", "CLIENT_TYPE"] },
...
      "Employee": { "type": "object", "properties": { "EmployeeName": { "type": "string" }, "Gender": { "type": "string" }, "BirthDate": { "type": "string" }, "Job_ID": { "type": "integer" }, "Address": { "type": "string" }, "Phone": { "type": "string" }, "Salary": { "type": "number" }, "Remarks": { "type": "string" }, "Photo": { "type": "string", "format": "binary" } }, "required": ["EmployeeName"] },
...
      "Invoice": { "type": "object", "properties": { "Client_no": { "type": "integer" }, "EmployeeID": { "type": "integer" }, "Invoice_status": { "type": "string" }, "InvoiceMemo": { "type": "string" }, "Invoice_date": { "type": "string", "format": "date" } }, "required": ["Client_no", "EmployeeID"] },
      "InvoiceDetail": { "type": "object", "properties": { "InvoiceNo": { "type": "integer" }, "Product_no": { "type": "string" }, "Qty": { "type": "number" }, "Price": { "type": "number" } }, "required": ["InvoiceNo", "Product_no", "Qty", "Price"] }
    }
  }
}